Move the executor and spawn package into driver
